prac7/prac7/customer-service/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
requests==2.31.0
stomp.py==8.1.0
